+++ vite.config.ts
@@ -64,8 +64,8 @@ export default defineConfig(() => {
     server: {
       port: 3000,
       host: '0.0.0.0',
+      strictPort: true,
       hmr: process.env.DISABLE_HMR !== 'true',
-      watch: process.env.DISABLE_HMR === 'true' ? null : {},
     },
     preview: {
       port: 3000,